moved proprietary types to driver-base
